requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
bcrypt==4.1.2
email-validator==2.1.0

# Testing dependencies (optional)
pytest==7.4.3
pytest-flask==1.3.0